bug fixes and debug strings
in write_buffer_queue need to set cur_write_buffer to NULL when sending
the cur_write_buffer buffer (happens when a frame ends exactly at the end
of the buffer)

+++ vod/common.h
@@ -169,7 +169,7 @@ enum {
 };
 
 typedef int bool_t;
-typedef int vod_status_t;
+typedef intptr_t vod_status_t;
 
 typedef struct {
 	vod_pool_t* pool;
